Work on Schedule Parsing
Schedule parsing for the regular season is finished. Need to work on
parsing for the playoffs though. Right now we are not counting playoffs.
We could add some more data structure and count playoffs fairly easily,
but then we have the issue of how to differentiate real playoffs from
consolation bracket.